Planning/src/requirement.txt
paho-mqtt==1.5.1
numpy==1.20.2
scipy==1.6.2
Flask==1.0.0
pandas==1.2.3
scikit_learn==0.24.1
PyYAML==5.4.1